chore: Avoid using `as` when casting from larger to smaller integer types

This MR addresses a concern from a ToB review:

> The `Governance::check_neuron_population_can_grow` method (in `sns/governance/src/governance.rs`) uses as to cast from `u64` to `usize` (which is 32 bits on 32-bit targets like `wasm32-unknown-unknown`). This should be avoided since this can cause truncation issues if the developer is not careful. 

See merge request dfinity-lab/public/ic!14348

+++ rs/sns/governance/src/governance.rs
@@ -5026,10 +5026,9 @@ impl Governance {
         let max_number_of_neurons = self
             .nervous_system_parameters_or_panic()
             .max_number_of_neurons
-            .expect("NervousSystemParameters must have max_number_of_neurons")
-            as usize;
+            .expect("NervousSystemParameters must have max_number_of_neurons");
 
-        if self.proto.neurons.len() + 1 > max_number_of_neurons {
+        if (self.proto.neurons.len() as u64) + 1 > max_number_of_neurons {
             return Err(GovernanceError::new_with_message(
                 ErrorType::PreconditionFailed,
                 "Cannot add neuron. Max number of neurons reached.",
